Optimise +/- by reducing clones & shifts.
Doing an additive operation (with a - b == a + (-b)) now requires 1 or 2
shifts (if the values have the same sign) or 1 to 3 (if the values have
different signs), which is much less. Also, the various ref forms can
skip cloning some arguments in some cases, e.g. x + &y doesn't need to
always clone y, as it did previously.

Fixes #2 (by removing the code).